thymeleaf-extras-springsecurity

+++ src/main/java/com/example/assignment01/models/Order.java
@@ -19,7 +19,7 @@ public class Order {
     private Date orderDate;
 
     @Column(name = "total_money")
-    private Double totalMoney;
+    private Float totalMoney;
 
     @ManyToOne
     private User user;
