+++ src/main/java/com/example/assignment01/models/Product.java
@@ -17,7 +17,7 @@ public class Product {
 
     private String name;
 
-    private Double price;
+    private Float price;
 
     private String description;
